Fix image/mask type before augmentations

+++ angionet/datasets/sennet.py
@@ -57,7 +57,7 @@ class TrainDataset(Dataset):
         D = 2 if self.dtms else 1
         masks = np.stack(sample[1:], dtype = 'float32')[:D, self.class_index]
         masks = chain.from_iterable([unbind(m) for m in unbind(masks)])
-        augs = self.transforms(image=sample[0].astype('float32'), masks=list(masks))
+        augs = self.transforms(image=sample[0], masks=list(masks))
 
         masks = [torch.stack(augs["masks"][: len(self.class_index)])]
         if self.dtms:
